Time: 7 ms (71.50%), Space: 8.4 MB (35.56%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -4,7 +4,7 @@ public:
     {                               
       int temp = x;  
       long rev = 0;
-      int digit;  
+      uint8_t digit;  
       while(temp > 0)  
       {
         digit = temp%10;
